+++ src/bench/crypto_hash.cpp
@@ -5,13 +5,12 @@
 #include <iostream>
 
 #include "bench.h"
-#include "bloom.h"
 #include "crypto/ripemd160.h"
 #include "crypto/sha1.h"
 #include "crypto/sha256.h"
 #include "crypto/sha512.h"
+#include "crypto/siphash.h"
 #include "random.h"
-#include "utiltime.h"
 
 /* Number of bytes to hash per iteration */
 static const uint64_t BUFFER_SIZE = 1000*1000;
