+++ src/coins.cpp
@@ -10,6 +10,7 @@
 #include "invalid.h"
 #include "logging.h"
 #include "random.h"
+#include "version.h"
 
 #include <assert.h>
