+++ src/test/hash_tests.cpp
@@ -2,6 +2,7 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
+#include "crypto/siphash.h"
 #include "hash.h"
 #include "utilstrencodings.h"
 #include "test/test_pivx.h"
