+++ src/txmempool.h
@@ -13,6 +13,7 @@
 
 #include "amount.h"
 #include "coins.h"
+#include "crypto/siphash.h"
 #include "indirectmap.h"
 #include "policy/feerate.h"
 #include "primitives/transaction.h"
